requirements.txt
GitPython>=3.1.41
jupyter>=1.0.0
mysql-connector-python>=8.3.0
pandas>=2.2.0
plotly>=5.18.0
streamlit>=1.30.0
streamlit-option-menu>=0.3.12
